feat: add security

+++ apps/frontend/features/judge/data/forms.ts
@@ -1,7 +1,9 @@
+"use server";
+
+import { verifyAction } from "@/features/auth/data/auth";
 import { CategoryFormSchemaType } from "../components/CategoryForm";
 import { JudgeFormMainSchema } from "../components/JudgeFormMain";
 import { JudgeFormOthersSchema } from "../components/JudgeFormOther";
-import { PhotoData } from "../types/main";
 
 export interface JudgeFormMainData {
   data: JudgeFormMainSchema;
@@ -17,11 +19,15 @@ export const onJudgeFormMainSubmit = async ({
   data,
   photoId,
 }: JudgeFormMainData) => {
+  const decoded = await verifyAction();
+  if (!decoded) throw new Error("Unauthorized access");
+
   const res = await fetch(
     `${process.env.NEXT_PUBLIC_API_URL}/api/judge/main/${photoId}`,
     {
       headers: {
         "Content-Type": "application/json",
+        Authorization: `Bearer ${decoded.token}`,
       },
       method: "POST",
       body: JSON.stringify({
@@ -40,11 +46,16 @@ export const onJudgeFormOtherSubmit = async ({
   data,
   photoId,
 }: JudgeFormOtherData) => {
+  const decoded = await verifyAction();
+
+  if (!decoded) throw new Error("Unauthorized access");
+
   const res = await fetch(
     `${process.env.NEXT_PUBLIC_API_URL}/api/judge/other/${photoId}`,
     {
       headers: {
         "Content-Type": "application/json",
+        Authorization: `Bearer ${decoded.token}`,
       },
       method: "POST",
       body: JSON.stringify({
@@ -64,11 +75,15 @@ export const onAssignAwardSubmit = async ({
 }: {
   awards: { categoryId: string; photoId: string | undefined }[];
 }) => {
+  const decoded = await verifyAction();
+  if (!decoded) throw new Error("Unauthorized access");
+
   const res = await fetch(
     `${process.env.NEXT_PUBLIC_API_URL}/api/judge/assign-award`,
     {
       headers: {
         "Content-Type": "application/json",
+        Authorization: `Bearer ${decoded.token}`,
       },
       method: "POST",
       body: JSON.stringify({
@@ -83,9 +98,13 @@ export const onAssignAwardSubmit = async ({
 };
 
 export const onCreateCategorySubmit = async (data: CategoryFormSchemaType) => {
+  const decoded = await verifyAction();
+  if (!decoded) throw new Error("Unauthorized access");
+
   const res = await fetch(`${process.env.NEXT_PUBLIC_API_URL}/api/categories`, {
     headers: {
       "Content-Type": "application/json",
+      Authorization: `Bearer ${decoded.token}`,
     },
     method: "POST",
     body: JSON.stringify(data),
@@ -100,11 +119,16 @@ export const onUpdateCategorySubmit = async (
   data: CategoryFormSchemaType,
   categoryId: string,
 ) => {
+  const decoded = await verifyAction();
+
+  if (!decoded) throw new Error("Unauthorized access");
+
   const res = await fetch(
     `${process.env.NEXT_PUBLIC_API_URL}/api/categories/${categoryId}`,
     {
       headers: {
         "Content-Type": "application/json",
+        Authorization: `Bearer ${decoded.token}`,
       },
       method: "PUT",
       body: JSON.stringify(data),
@@ -117,10 +141,17 @@ export const onUpdateCategorySubmit = async (
 };
 
 export const onDeleteCategorySubmit = async (categoryId: string) => {
+  const decoded = await verifyAction();
+
+  if (!decoded) throw new Error("Unauthorized access");
+
   const res = await fetch(
     `${process.env.NEXT_PUBLIC_API_URL}/api/categories/${categoryId}`,
     {
       method: "DELETE",
+      headers: {
+        Authorization: `Bearer ${decoded.token}`,
+      },
     },
   );
